+++ src/OpenTelemetry.Instrumentation.SqlClient/CHANGELOG.md
@@ -2,6 +2,12 @@
 
 ## Unreleased
 
+* Use `Activity.Status` and `Activity.StatusDescription` properties instead of
+  `OpenTelemetry.Trace.Status` and `OpenTelemetry.Trace.Status.Description`
+  respectively to set activity status.
+  ([#3118](https://github.com/open-telemetry/opentelemetry-dotnet/issues/3118))
+  ([#3751](https://github.com/open-telemetry/opentelemetry-dotnet/pull/3751))
+
 ## 1.0.0-rc9.7
 
 Released 2022-Sep-29